+++ Services/Command Line/Implementation/CommandLineArgumentsService.swift
@@ -25,7 +25,11 @@ struct CommandLineArgumentsService: ArgumentsService {
         for i in 1 ..< CommandLine.arguments.count {
             let argumentString = CommandLine.arguments[i]
             if let argumentKey = Argument.Key(rawValue: argumentString) {
-                parsingMode = argumentKey
+                if argumentKey == .verbose {
+                    result.append(Argument(key: argumentKey, value: nil))
+                } else {
+                     parsingMode = argumentKey
+                }
             } else if let mode = parsingMode,
                 let argumentModel = mode.argument(value: argumentString) {
                 result.append(argumentModel)