+++ CHANGELOG.md
@@ -4,6 +4,14 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.0.0] - 2019-07-11
+### Added
+- Return code emitted to fail CI jobs where upload fails.
+- `--verbose` flag added in order to obtain more detailed output.
+
+### Changed
+- Output made more concise.
+
 ## [0.0.2] - 2018-11-26
 ### Added
 - Ability to specify a timeout in seconds for the upload using the `--timeout` parameter.